tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-svg-rect-manual.html

<!DOCTYPE html>
<title>Element#requestFullscreen() for SVG rect element</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<svg><rect/></svg>
<script>
async_test(function(t)
{
    var rect = document.querySelector("rect");
    assert_true(rect instanceof SVGRectElement);
    trusted_request(t, rect, document.body);
    document.onfullscreenchange = t.unreached_func("fullscreenerror event");
    document.onfullscreenerror = t.step_func_done();
});
</script>
